fix: remaining sonarcube code dupe issues
